+++ Common/src/main/java/com/tristankechlo/wool_collection/platform/IPlatformHelper.java
@@ -1,12 +1,16 @@
 package com.tristankechlo.wool_collection.platform;
 
+import com.tristankechlo.wool_collection.TheWoolCollection;
 import com.tristankechlo.wool_collection.container.WeavingStationContainer;
 import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.CreativeModeTab;
 
 import java.nio.file.Path;
 import java.util.function.Supplier;
 
 public interface IPlatformHelper {
+
+    public static final IPlatformHelper INSTANCE = TheWoolCollection.load(IPlatformHelper.class);
 
     String getPlatformName();
 
@@ -22,5 +26,7 @@ public interface IPlatformHelper {
     Path getConfigDirectory();
 
     Supplier<MenuType<WeavingStationContainer>> buildContainer();
+
+    CreativeModeTab.Builder getCreativeTabBuilder();
 
 }