+++ Fabric/src/main/java/com/tristankechlo/wool_collection/TheWoolCollectionFabric.java
@@ -3,11 +3,9 @@ package com.tristankechlo.wool_collection;
 import com.tristankechlo.wool_collection.init.ModBlocks;
 import com.tristankechlo.wool_collection.init.ModRegistry;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.fabricmc.fabric.api.registry.FlammableBlockRegistry;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.level.block.state.properties.BlockSetType;
 import net.minecraft.world.level.block.state.properties.WoodType;
 
@@ -28,24 +26,6 @@ public class TheWoolCollectionFabric implements ModInitializer {
         //register all items
         ModBlocks.ALL_ITEMS.forEach((id, item) -> {
             Registry.register(BuiltInRegistries.ITEM, id, item);
-        });
-
-        // add all items to creative tab for colored blocks
-        ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.COLORED_BLOCKS).register((content) -> {
-            TheWoolCollection.sortedListByColor(ModBlocks.FENCES).forEach(content::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.FENCE_GATES).forEach(content::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.STAIRS).forEach(content::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.SLABS).forEach(content::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.WALLS).forEach(content::accept);
-        });
-        // add all buttons to redstone creative tab
-        ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.REDSTONE_BLOCKS).register((content) -> {
-            TheWoolCollection.sortedListByColor(ModBlocks.BUTTONS).forEach(content::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.PRESSURE_PLATES).forEach(content::accept);
-        });
-        // add all items to functional blocks creative tab
-        ItemGroupEvents.modifyEntriesEvent(CreativeModeTabs.FUNCTIONAL_BLOCKS).register((content) -> {
-            content.accept(ModRegistry.WEAVING_STATION_ITEM.get());
         });
     }