+++ Fabric/src/main/java/com/tristankechlo/wool_collection/platform/FabricPlatformHelper.java
@@ -1,9 +1,11 @@
 package com.tristankechlo.wool_collection.platform;
 
 import com.tristankechlo.wool_collection.container.WeavingStationContainer;
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
 import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.world.flag.FeatureFlags;
 import net.minecraft.world.inventory.MenuType;
+import net.minecraft.world.item.CreativeModeTab;
 
 import java.nio.file.Path;
 import java.util.function.Supplier;
@@ -33,6 +35,11 @@ public class FabricPlatformHelper implements IPlatformHelper {
     @Override
     public Supplier<MenuType<WeavingStationContainer>> buildContainer() {
         return () -> new MenuType<>(WeavingStationContainer::new, FeatureFlags.VANILLA_SET);
+    }
+
+    @Override
+    public CreativeModeTab.Builder getCreativeTabBuilder() {
+        return FabricItemGroup.builder();
     }
 
 }