+++ Forge/src/main/java/com/tristankechlo/wool_collection/TheWoolCollectionForge.java
@@ -2,12 +2,10 @@ package com.tristankechlo.wool_collection;
 
 import com.tristankechlo.wool_collection.init.ModBlocks;
 import com.tristankechlo.wool_collection.init.ModRegistry;
-import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.FireBlock;
 import net.minecraft.world.level.block.state.properties.BlockSetType;
 import net.minecraft.world.level.block.state.properties.WoodType;
-import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
@@ -26,7 +24,6 @@ public class TheWoolCollectionForge {
         IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
         modEventBus.addListener(this::onRegister);
         modEventBus.addListener(this::onCommonSetup);
-        modEventBus.addListener(this::onCreativeModeTabRegister);
     }
 
     /* register all blocks and items */
@@ -36,24 +33,6 @@ public class TheWoolCollectionForge {
         }
         if (event.getRegistryKey() == ForgeRegistries.Keys.ITEMS) {
             ModBlocks.ALL_ITEMS.forEach((id, item) -> event.register(ForgeRegistries.Keys.ITEMS, id, () -> item));
-        }
-    }
-
-    /* add items to their creative tab */
-    private void onCreativeModeTabRegister(BuildCreativeModeTabContentsEvent event) {
-        if (event.getTabKey() == CreativeModeTabs.COLORED_BLOCKS) {
-            TheWoolCollection.sortedListByColor(ModBlocks.FENCES).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.FENCE_GATES).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.STAIRS).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.SLABS).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.WALLS).forEach(event::accept);
-        }
-        if (event.getTabKey() == CreativeModeTabs.REDSTONE_BLOCKS) {
-            TheWoolCollection.sortedListByColor(ModBlocks.BUTTONS).forEach(event::accept);
-            TheWoolCollection.sortedListByColor(ModBlocks.PRESSURE_PLATES).forEach(event::accept);
-        }
-        if (event.getTabKey() == CreativeModeTabs.FUNCTIONAL_BLOCKS) {
-            event.accept(ModRegistry.WEAVING_STATION_ITEM.get());
         }
     }